requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-google
gspread
gspread-dataframe
Authlib
boto3